requirements.txt
nltk==3.2.4
tensorflow-gpu~=1.15.0